backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.0
requests>=2.27.0
mailchimp3==3.0.1
tweepy==4.14.0
linkedin-api==2.0.0a5
gunicorn==21.2.0 